Buttons work properly... or almost
To PlayerAction.cs it has been added a Debug.Log to warn if the action
called is inactive and also an extra parameter to PlayerAction(...), so
an action can start active or not.

Still, the Actions doesn't know yet which button they should make
inactive, so calling actionUsed() results in an error.