Bug #2 solved
ActionManager.Start() was properly constructing all the PlayerActions...

...but the values used were being initialized in StatsManager.cs, inside
a Start(). Because of that, the StatsManager.Start() would be called
AFTER ActionManager.Start(), and the numeric change of the first
PlayerAction created would be always equal to 0.

Fault solved through passing the code in StatsManager.Start() to
StatsManager.Awake(). Since Awake() is called before all the Start(),
the values called in ActionManager.Start() are all properly initialized
at the right time.