(Fixing #7) Missing digits solved enough
Now there are more digits being shown both for MonthlyBudget and
CurrentMoney.

There are a few problems yet of how many digits is acceptable to show
and not showing the decimal point, but such things are better solved
while we balance the simulation, to know exactly what to show and what
not.