(Fixed #12) Goodbye InvalidOperationException
The fault that created this exception was this:

Foreach in this case used a collection, and the collection was being
changed in the middle of the foreach by its intended purpose: removing
objects from the collection when they became legal.

Solved by using a for without incrementing directly it's own counter:
the for doesn't need a static collection and with increments tied to the
absence of a change, no danger of ignoring any object.