(Working in #14) Fixing the Simulation
I'm starting to create the game simulation in Machinations, so we will
be able to see easier what is wrong in it and ways to fix it.

I found at least one thing that's wrong: since the houses in the game
now only apply to the refugees, it shouldn't influence criminality.