(#8) amount_of_refugees is public and multiplied
With this, we now have a value to start tinkering with for balancing the
game in the future.

I've also (shouldn't have done this in the same commit, sorry) changed
the equations for how many refugees there will be in a wave: now it's
default_amount_of_refugees times a random generated number, accordingly
to the situation.